bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
click==8.1.8
discord.py==2.5.2
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.6.0
h11==0.16.0
idna==3.10
multidict==6.4.3
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
yarl==1.20.0
